Read from file

+++ munro-lib/src/main/java/com/example/munro/Munro.kt
@@ -1,6 +1,6 @@
 package com.example.munro
 
-data class MunroResult(val name: String, val heightMetre: Int, val hillCategory: MunType, val gridReference: String): Comparable<MunroResult> {
+data class MunroResult(val name: String, val heightMetre: Double, val hillCategory: MunType, val gridReference: String): Comparable<MunroResult> {
     override fun compareTo(other: MunroResult): Int {
         return heightMetre.compareTo(other.heightMetre)
     }
